# docs/features/F-4-branches/openapi.yaml
openapi: 3.1.1
info:
  title: F-4 Ветки API
  version: 1.0.0
  description: Контракт управления ветками текущего пользователя.
servers:
  - url: /
security:
  - bearerAuth: []
tags:
  - name: Branches
paths:
  /api/v1/branches:
    get:
      tags: [Branches]
      operationId: getBranches
      summary: Список и поиск активных веток
      description: |
        Только ветки пользователя, определённого Gateway по проверенному access token. Сортировка по createdAt, затем id по убыванию.
        Поиск по части названия без учёта регистра; %, _ и другие символы запроса
        трактуются буквально. Пустой после нормализации q равен отсутствующему фильтру.
        cursor привязан к нормализованному q с сохранением регистра. Вставки до текущего курсора видны
        после обновления списка с первой страницы.
      parameters:
        - name: limit
          in: query
          description: Число элементов на странице.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
        - name: cursor
          in: query
          description: Непрозрачный nextCursor из предыдущего ответа.
          schema:
            type: string
            maxLength: 512
        - name: q
          in: query
          description: Подстрока названия; после NFC и удаления краевых пробелов не более 100 кодовых точек Unicode. Правило нормализации — в system-description.md, раздел 4.
          schema:
            type: string
            x-normalizedMaxCodePoints: 100
      responses:
        '200':
          description: Страница активных веток.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
    post:
      tags: [Branches]
      operationId: createBranch
      summary: Создать ветку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchInput'
      responses:
        '201':
          description: Ветка создана.
          headers:
            Location:
              description: Путь к созданной ветке.
              schema:
                type: string
                example: /api/v1/branches/8f20d971-bdd8-4d2b-82ea-5fd4e126532f
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
  /api/v1/branches/{branchId}:
    parameters:
      - $ref: '#/components/parameters/BranchId'
    get:
      tags: [Branches]
      operationId: getBranch
      summary: Получить активную ветку
      responses:
        '200':
          description: Активная ветка.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
    put:
      tags: [Branches]
      operationId: updateBranch
      summary: Переименовать активную ветку
      description: Если нормализованное название не изменилось, updatedAt и аудит не меняются.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchInput'
      responses:
        '200':
          description: Актуальная ветка.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
  /api/v1/branches/{branchId}/archive:
    parameters:
      - $ref: '#/components/parameters/BranchId'
    post:
      tags: [Branches]
      operationId: archiveBranch
      summary: Архивировать собственную ветку
      description: |
        Повторное архивирование собственной ветки возвращает 204 без изменения
        archivedAt и без нового события аудита. Чужая и отсутствующая ветка — 404.
        В F-4 архивируется только ветка; F-5/F-6 добавят каскадное архивирование.
      responses:
        '204':
          description: Ветка архивирована или уже была архивирована.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Для маршрутов веток Gateway проверяет подписанный access token с type=access, корректными iss, exp и UUID в sub. Tasks получает подписанный внутренний контекст без пользовательского токена.
  parameters:
    BranchId:
      name: branchId
      in: path
      required: true
      description: UUID ветки. Некорректный формат — 400.
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Ошибка валидации входных данных.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Проверьте введённые данные
              details:
                - field: name
                  message: Название должно содержать от 2 до 100 символов
    Unauthorized:
      description: Gateway обнаружил отсутствие или недействительность access token до вызова tasks.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Требуется авторизация
              details: null
    NotFound:
      description: Ветка не найдена или недоступна.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Ветка не найдена или недоступна
              details: null
    UpstreamUnavailable:
      description: После успешной проверки JWT сервис tasks отклонил внутренний контекст; Gateway не раскрывает внутреннюю причину клиенту.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UPSTREAM_UNAVAILABLE
              message: Сервис временно недоступен. Попробуйте позже
              details: null
  schemas:
    BranchInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          x-normalizedMinCodePoints: 2
          x-normalizedMaxCodePoints: 100
          description: После NFC и удаления краевых пробелов — 2–100 кодовых точек Unicode. minLength относится к исходной строке; семантические границы задают x-normalizedMinCodePoints и x-normalizedMaxCodePoints. Правило нормализации — в system-description.md, раздел 4.
          example: Спорт
    Branch:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BranchPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Branch'
        nextCursor:
          type: [string, 'null']
          maxLength: 512
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, NOT_FOUND, UPSTREAM_UNAVAILABLE]
            message:
              type: string
            details:
              type: [array, 'null']
              items:
                $ref: '#/components/schemas/ErrorDetail'